MCHEM: Switch to KPP version 3.1.0
This commit changes the default KPP version for MCHEM to 3.1.0 in order
to utilize the new autoreduce capabilities of KPP. Currently, the
autoreduce integrator setting is only used for the gas phase mechanism.

We will need to update the code for KPP version 3.2.0 in order to be in
line with variable concentrations in Update_Rconst().

Further, this commit includes minor changes to the aqueous phase
mechanism.