Allow compilation of combine with gcc10+ This code change addresses issue #1150 https://github.com/USEPA/CMAQ_Dev/issues/1150
Function WRF_DESC in module_file.F previously contained
two calls to NF_GET_VARA_REAL, with VGTOP3D specified
as last argument in the first call and VGLVLS3D specified
in the second call. The rank mismatch between the first
call (scalar) and the second call (array) triggered
an error when compiling the code with gcc10+ as
documented in the issue description.

After reviewing the code, it was determined that the
vertical coordinate information of the wrfout file
generated by the WRF_DESC function, i.e. VGTYP3D (which
was somehow hardcoded as 7), VGTOP3D, and VGLVLS3D
is not actually being used for any consistency
checking between the wrfout vertical structure and the
vertical structure of INFILE1, i.e. the first I/O API
opened by combine that is used to set the grid structure
of OUTFILE (in contrast to the wrfout horizontal grid
and timestep information also obtained by WRF_DESC, which
is being checked for consistency with that of INFILE1).

Given that the rank-mismatched VGTOP3D and VGLVLS3D
valued obtained by the successive calls to NF_GET_VARA_REAL
are not used in any code and cause a problem for gcc10+,
the issue was addressed by removing these calls, as well
as the unnecessary hardcoding of VGTYP3D.

	modified:   module_file.F